Don't apply floors in robust inverter
The Kastaun inverter is guaranteed to converge with only a tiny
minimum of limits.  This commit restricts the implementation
to apply only these.

Without internal floors, results can be very unstable, especially
for the case that no velocity will satisfy the momentum; for this,
the inverter reconstructs the maximum possible velocity.

Thus we apply an adapted form of Normal Observer floors after
inversion, which are guaranteed to add at least enough
material and energy to satisfy constraints.  We also check that
the new primitives have the correct momentum -- if not, the
velocity is zeroed to avoid myriad problems evolving high speeds.
Most of the momentum in these cells comes from B, not rho,
so I think this is a reasonable choice in practice.